Refactor, move device definitions to separate folder.
